chore(archtecture): rename folder
