+++ src/infra/tattoos/tattoo.gateway.ts
@@ -1,4 +1,4 @@
-import { Gateway } from "../architecture/gateway/Gateway";
+import { Gateway } from "../core/gateway/Gateway";
 import { Data, ITattoo } from "./tattoo.interface";
 import { ParamsDTO } from "./tattoo.type";
