+++ src/infra/users/users.gateway.ts
@@ -1,4 +1,4 @@
-import { Gateway } from "../architecture/gateway/Gateway";
+import { Gateway } from "../core/gateway/Gateway";
 import { Data } from "../tattoos/tattoo.interface";
 import { User } from "./user.type";
 import { InfoMe } from "./users.interface";
